feature: Remove the missing arch feature in event.rs.
