+++ src/arch/riscv64/ipi.rs
@@ -13,6 +13,7 @@
 //
 // Authors:
 //
+use crate::consts::{IPI_EVENT_SEND_IPI, IPI_EVENT_UPDATE_HART_LINE};
 use sbi_rt::HartMask;
 
 // arch_send_event
@@ -29,4 +30,25 @@ pub fn arch_ipi_handler() {
     unsafe {
         riscv_h::register::hvip::set_vssip();
     }
+}
+
+pub fn arch_check_events(event: Option<usize>) {
+    match event {
+        #[cfg(feature = "plic")]
+        Some(IPI_EVENT_UPDATE_HART_LINE) => {
+            use crate::device::irqchip::plic::update_hart_line;
+            update_hart_line();
+        }
+        Some(IPI_EVENT_SEND_IPI) => {
+            use crate::arch::riscv64::ipi::arch_ipi_handler;
+            arch_ipi_handler();
+        }
+        _ => {
+            panic!("arch_check_events: unhandled event: {:?}", event);
+        }
+    }
+}
+
+pub fn arch_prepare_send_event() {
+    debug!("risc-v arch_prepare_send_event: do nothing now.")
 }
